Move entry type method tests to basic-test.test.ts
- Consolidate tests into basic-test.test.ts to match existing test structure
- Remove console.log statements to match existing test style
- Simplify test names and variable names to match existing patterns
- Test entry type methods with hyphenated, camelCase, and underscore naming
